+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Added data import failure detection for when it's not possible to deduce the CSV delimiter
   * Added special case for handling the dataLabels type
   * Added support for the `series-label` module
+  * Google Spreadsheet support now extracts the correct ID from a URL
 
 # 0.2.1-rc2
   * Fixed issue causing propagation of importer properties to fail